backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
openai
sqlalchemy[asyncio]
asyncpg
alembic
slowapi
PyJWT
bcrypt
python-multipart
# test dependencies
pytest
pytest-asyncio
httpx
aiosqlite
slowapi